+++ GUI/Controls/GUIRadioButton.cpp
@@ -82,7 +82,7 @@ namespace RTE {
 
 	void GUIRadioButton::BuildBitmap() {
 		std::string Filename;
-		unsigned long ColorIndex = 0;
+		int ColorIndex = 0;
 		int Values[4];
 
 		// Load the image
